backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
bcrypt==4.3.0
python-dotenv==1.0.1
pydantic-settings==2.7.1
pydantic[email]==2.10.5
pdfplumber==0.11.4
PyMuPDF==1.25.3
python-docx==1.1.2
openpyxl==3.1.5
langchain==0.3.14
langchain-groq==0.2.4
fpdf2==2.8.2
httpx==0.28.1
Pillow>=10.0.0
google-generativeai>=0.8.0

